+++ equalsverifier-core/src/main/java/nl/jqno/equalsverifier/internal/checkers/fieldchecks/NullPointerExceptionFieldCheck.java
@@ -41,9 +41,13 @@ public class NullPointerExceptionFieldCheck<T> implements FieldCheck<T> {
             Object saved = fieldProbe.getValue(reference);
 
             // Safe, because we already checked the field is non-final
-            fieldMutator.setNewValue(reference, PrimitiveMappers.DEFAULT_VALUE_MAPPER.get(fieldProbe.getType()));
-            performTests(fieldProbe.getField(), subjectCreator.plain(), subjectCreator.plain());
-            fieldMutator.setNewValue(reference, saved);
+            try {
+                fieldMutator.setNewValue(reference, PrimitiveMappers.DEFAULT_VALUE_MAPPER.get(fieldProbe.getType()));
+                performTests(fieldProbe.getField(), subjectCreator.plain(), subjectCreator.plain());
+            }
+            finally {
+                fieldMutator.setNewValue(reference, saved);
+            }
         }
         else {
             T changed = subjectCreator.withFieldDefaulted(fieldProbe.getField());